refactor/fix: defined ScheduleRecoveryLessonButton component to solve random save recovery lesson bug
